requirements.txt
os
request
bs4
streamlit
datetime
urllib.parse
langchain
langchain_core
langchain_community
langchain_groq
python_dotenv
gtts